% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashr.R
\name{cash}
\alias{cash}
\title{Fit Empirical Bayes Normal Means with Correlated Noise}
\usage{
cash(x, s = 1, deltaAt0 = TRUE, gd.order = 10, omega.lambda = 10,
  omega.rho = 0.5, omega.pen = NULL, mixsd.mult = sqrt(2),
  gd.priority = FALSE, control = list(maxiter = 50))
}
\arguments{
\item{x}{A p vector of observations}

\item{s}{A scalar or a p vector of standard deviations.}

\item{deltaAt0}{Logical, indicating whether to use a point mass at zero as one of components for a mixture distribution of the prior.}

\item{gd.priority}{Logical, indicating whether to optimizer over prior}
}
\description{
This is the main interface for fitting correlated EBNM models
  based on algorithms proposed by Sun and Stephens.  The default
  behaviour is simply to run the biconvex optimization and return
  the result and posterior calculations.
}
