% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecn.R
\name{ecn}
\alias{ecn}
\title{Fit empirical distribution of observed correlated N(0,1) noise using Exchangeable Correlated Noise (ECN) model}
\usage{
ecn(z, gd.order = 10, omega.lambda = 10, omega.rho = 0.5,
  omega.pen = NULL)
}
\description{
Fit empirical distribution of observed correlated N(0,1) noise using Exchangeable Correlated Noise (ECN) model
}
