% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecn.R
\name{std.gaussDeriv}
\alias{std.gaussDeriv}
\title{Calculate the values of standardized Gaussian derivatives at a given set of positions}
\usage{
std.gaussDeriv(x, gd.order)
}
\arguments{
\item{gd.order}{}
}
\description{
Calculate the values of standardized Gaussian derivatives at a given set of positions
}
